fix: Fix the search function is not available (MC 1.16)
- No doubt blame LibGui
