--- src/main/java/cn/msdnicrosoft/commandbuttons/gui/CommandEntryTextField.java
@@ -1,24 +1,0 @@
-package cn.msdnicrosoft.commandbuttons.gui;
-
-import io.github.cottonmc.cotton.gui.widget.WTextField;
-import lombok.Setter;
-
-import java.util.function.Consumer;
-
-public class CommandEntryTextField extends WTextField {
-    @Setter
-    private Consumer<String> focusLostCallback;
-
-    @Override
-    public void releaseFocus() {
-        super.releaseFocus();
-    }
-
-    @Override
-    public void onFocusLost() {
-        if (this.focusLostCallback != null) {
-            focusLostCallback.accept(this.getText());
-        }
-        super.onFocusLost();
-    }
-}
